[Feature]: Code splitting
